Move non-entrypoint JavaScript into src/
While there may have been some rhyme or reason for creating these files
outside of src/ originally, it's no longer apparent.  It's nice to
systematize the file organization a bit, esp. in advance of additional
files I'm planning to add to support new server-side API endpoints.
