+++ README.md
@@ -79,7 +79,7 @@ Nextstrain documentation is hosted by Read The Docs at [docs.nextstrain.org](htt
 Please see [this GitHub repo](https://github.com/nextstrain/docs.nextstrain.org/) for more details.
 
 > Note that the documentation used to be served from the server in this repo at URLs such as nextstrain.org/docs/... until [November 2020](https://github.com/nextstrain/nextstrain.org/pull/226).
-A number of [redirects](./redirects.js) have been added to preserve old URLs.
+A number of [redirects](./src/redirects.js) have been added to preserve old URLs.
 
 ---
 ## Auspice client
