+++ docs/infrastructure.md
@@ -26,7 +26,7 @@ Deploys of `master` happen automatically after Travis CI tests are successful.
 
   - `REDIS_URL` is provided by the Heroku Redis add-on.
     It should not be modified directly.
-    Our [authentication handlers](../authn.js) rewrite it at server start to use a secure TLS connection.
+    Our [authentication handlers](../src/authn/index.js) rewrite it at server start to use a secure TLS connection.
 
   - `FETCH_CACHE` is not currently used, but can be set to change the location of the on-disk cache used by (some) server `fetch()`-es.
     The default location is `/tmp/fetch-cache`.
@@ -107,7 +107,7 @@ Ephemeral instances are automatically managed by AWS Batch for `nextstrain build
 ## Cognito
 
 A [_user pool_](https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-identity-pools.html) called `nextstrain.org` provides authentication for Nextstrain logins.
-Cognito is integrated with the nextstrain.org server using the OAuth2 support from PassportJS in our [`authn.js`](../authn.js) file.
+Cognito is integrated with the nextstrain.org server using the OAuth2 support from PassportJS in our [`authn/index.js`](../src/authn/index.js) file.
 
 We currently don't use Cognito's [_identity pools_](https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-identity.html).
 It may be beneficial to use one in the future so we can get temporary AWS credentials specific to each Nextstrain user with the appropriate authorizations baked in (instead of using a server-wide set of credentials).
